// client/src/models/moviedb/tv-show-details.model.ts
import CreatedBy from "./created-by.model";
import Genre from "./genre.model";
import LastEpisodeToAir from "./last-episode-to-air.model";
import Network from "./network.model";
import ProductionCompany from "./production-company.model";
import Season from "./season.model";

export default interface TvShowDetails {
  backdrop_path: string;
  created_by: CreatedBy[];
  episode_run_time: number[];
  first_air_date: string;
  genres: Genre[];
  homepage: string;
  id: number;
  in_production: boolean;
  languages: string[];
  last_air_date: string;
  last_episode_to_air: LastEpisodeToAir;
  name: string;
  next_episode_to_air?: any;
  networks: Network[];
  number_of_episodes: number;
  number_of_seasons: number;
  origin_country: string[];
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: ProductionCompany[];
  seasons: Season[];
  status: string;
  type: string;
  vote_average: number;
  vote_count: number;
}